==> data_analysis_python/01_parse_metadata_json_extract_boots.ipynb <==
import json
from pandas import *
import os
import sys
%load_ext autotime

f = open('meta_Clothing_Shoes_and_Jewelry.json').readlines()

keys = ["'asin':"
, "'title':"
, "'price':"
, "'imUrl':"
, "'related':"
, "'salesRank':"
, "'brand':"
, "'categories':"
, "'description':"]

def extract_key_value(kv, item_dict):
    kv_split = kv.split(": ")
    k = kv_split[0][1:].replace("'",'')
    v = ": ".join(kv_split[1:]).replace("''","','")
    if v[0] == "'":
        v = v[1:-1]
    item_dict[k] = v
    return item_dict

def make_item_dict(item):
    item_dict = {}
    curr = ''

    for i in item:
        key_found = [k for k in keys if k in i]
        if len(key_found) == 1:
            if curr != '':
                item_dict = extract_key_value(curr, item_dict)
            curr = i
        elif len(key_found) == 0:
            curr += i
        else:
            print 'TOO MANY KEYS'
            print i
            print item
            sys.exit(0)
    
    return item_dict

df = DataFrame.from_dict({i: make_item_dict(l[1:-2].split(', ')) for i, l in enumerate(f)}, orient = 'index')

df.to_csv('metadata.csv')

import tarfile
tar = tarfile.open("pics.tgz")
# pics = tar.getmembers()

pics_names_all = [p.replace('pics/','').replace('.jpg','') for p in tar.getnames() if '.jpg' in p]

df.loc[:, 'pic_downloaded'] = df.loc[:, "asin'"].apply(lambda x: x in pics_names_all)

df.head()

df.loc[:, 'boots'] = df.loc[:, "title'"].apply(lambda x: False if type(x) is float else 'boots' in x.lower())

df.to_csv('metadata_with_others.csv')

df.boots.value_counts()

df.pic_downloaded.value_counts()

boots = df[df.boots == True]

boots.head()

boots.to_csv('boots.csv')

boots_pics = boots[boots.boots == True].loc[:, "asin'"].values

boots_pics[:5]

[tar.extract('pics/{0}.jpg'.format(b)) for b in boots_pics]

### Picking back up
