==> data_analysis_python/05_clean_data.ipynb <==
from __future__ import division
from pandas import *
import math
import matplotlib.pyplot as plt
import matplotlib.image as mpimg
import os
import scipy


plt.style.use('ggplot')

%matplotlib inline
%load_ext autotime

data_path = '../boots_dataset/'

### load features -- drop all the duplicate images

features = read_csv(data_path + 'boots_features.csv', 
                    index_col = 'asin')\
                .drop_duplicates()
features.shape

#### save those back

features.to_csv(data_path + 'boots_features_dedup.csv')

features.head()

### load metadata

metadata = read_csv(data_path + 'boots_aws.csv', index_col = 'Unnamed: 0')
metadata = metadata[metadata.pic_downloaded == True]

metadata.head()

#### add a category of rain/snow boots

metadata.sex.value_counts()

def sex_of_boots(x):
    for rain_word in ['rain','snow']:
        if rain_word in x:
            return 'weather'        
    for woman_word in ['women', 'woman', 'ladies', 'lady', 'girl']:
        if woman_word in x:
            return 'women'
    for man_word in ['men', 'man', 'boy']:
        if man_word in x:
            return 'men'
    for child_word in ['kid', 'child', 'youth', 'baby', 'infant']:
        if child_word in x:
            return 'children'
    for unisex_word in ['unisex', 'adult']:
        if unisex_word in x:
            return 'unisex'
    else:
        return 'unknown'

metadata.loc[:, 'sex'] = metadata.loc[:, 'title']\
                .apply(lambda x: sex_of_boots(x.lower()))

metadata.sex.value_counts()

### make a men's/women's metadata set

metadata.to_csv(data_path + 'boots_aws_160403.csv')

men = metadata[metadata.sex == 'men']
women = metadata[metadata.sex == 'women']

men.to_csv(data_path + 'metadata_men.csv')
women.to_csv(data_path + 'metadata_women.csv')

### make a men's/women's feature set

features_women = features[features.index.isin(women.asin)]
features_men = features[features.index.isin(men.asin)]
features_women.to_csv(data_path + 'features_women.csv')
features_men.to_csv(data_path + 'features_men.csv')
