==> data_analysis_python/05b_clean_data_more.ipynb <==
from pandas import *
%load_ext autotime

data_path = '../boots_dataset/'

metadata = read_csv(data_path + 'boots_aws_160403.csv', index_col = 'Unnamed: 0').drop_duplicates()
metadata.index.name = '#'

metadata.head()

keys = ['also_viewed','also_bought','bought_together','buy_after_viewing']

def extract_key_value(kv, item_dict):
    kv_split = kv.split(": ")
    k = kv_split[0][1:].replace("'",'')
    v = ": ".join(kv_split[1:]).replace("''","','")
    if v[0] == "'":
        v = v[1:-1]
    item_dict[k] = v
    return item_dict

def make_item_dict(item):
    item_dict = {}
    curr = ''

    for i in item:
        key_found = [k for k in keys if k in i]
        if len(key_found) == 1:
            if curr != '':
                item_dict = extract_key_value(curr, item_dict)
            curr = i
        elif len(key_found) == 0:
            curr += i
        else:
            print 'TOO MANY KEYS'
            print i
            print item
            sys.exit(0)
    if curr != '':
        item_dict = extract_key_value(curr, item_dict)
    return item_dict

def extract_related(x):
    if type(x) is float:
        return x
    x = x[1:-2].replace("']'","'], '").split(', ')
    return make_item_dict(x)
    
metadata.loc[:, 'related_d'] = metadata.loc[:, 'related'].apply(lambda x: extract_related(x))

def extract_dict(x, k):
    if type(x) is float:
        return x
    if k in x.keys():
        return x[k].replace('[','').replace(']','').split(',')
    else:
        return float('nan')

for k in keys:
    metadata.loc[:, k] = metadata.loc[:, 'related_d']\
                                 .apply(lambda x: extract_dict(x, k))

metadata = metadata.drop('related_d', axis = 1)

metadata.head()

def sales_rank(x, val):
    if type(x) is float or x == '{}':
        return float('nan')
    x = x.replace('{','').replace('}','').replace("'",'').split(':')
    if val == 'cat':
        return x[0].strip()
    else:
        return x[1].strip()

metadata.loc[:, 'salesRank_category'] = metadata.loc[:, 'salesRank'].apply(lambda x: sales_rank(x, 'cat'))
metadata.loc[:, 'salesRank_rank'] = metadata.loc[:, 'salesRank'].apply(lambda x: sales_rank(x, 'rank'))

metadata.to_csv(data_path + 'boots_aws_metadata_160405.csv')
